+++ app/src/main/java/com/example/playlistmaker/common/data/ThemeRepository.kt
@@ -1,29 +1,6 @@
 package com.example.playlistmaker.common.data
 
-import android.content.Context
-import android.content.res.Configuration
-import android.util.Log
-import com.example.playlistmaker.main.domain.ThemeUseCase
-
-
-class ThemeRepository(private val context: Context) : ThemeUseCase {
-    companion object {
-        const val PREFS = "my_prefs"
-        const val THEME_PREF = "isDarkTheme"
-    }
-
-    private val prefs = context.getSharedPreferences(PREFS, Context.MODE_PRIVATE)
-
-    override fun isDarkTheme(): Boolean {
-        val currentNightMode = context.resources.configuration.uiMode and Configuration.UI_MODE_NIGHT_MASK
-        return prefs.getBoolean(THEME_PREF, currentNightMode == Configuration.UI_MODE_NIGHT_YES)
-    }
-
-    fun switchTheme(isDark: Boolean) {
-        with(prefs.edit()) {
-            putBoolean(THEME_PREF, isDark)
-            apply()
-        }
-        Log.d("ThemeRepository", "Theme switched to: ${if (isDark) "Dark" else "Light"}")
-    }
+interface ThemeRepository {
+    fun isDarkTheme(): Boolean
+    fun saveTheme(isDark: Boolean)
 }
