+++ app/src/main/java/com/example/playlistmaker/main/ui/activity/MainActivity.kt
@@ -6,7 +6,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.ViewModelProvider
 import com.example.playlistmaker.R
 import com.example.playlistmaker.common.data.ThemeRepository
+import com.example.playlistmaker.common.data.ThemeRepositoryImpl
 import com.example.playlistmaker.main.ui.view_model.MainViewModel
+import com.example.playlistmaker.settings.domain.ThemeUseCaseImpl
 
 class MainActivity : AppCompatActivity() {
 
@@ -15,8 +17,9 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        val themeRepository = ThemeRepository(applicationContext)
-        val factory = MainViewModel.getViewModelFactory(application, themeRepository)
+        val themeRepository: ThemeRepository = ThemeRepositoryImpl(applicationContext)
+        val themeUseCaseImpl = ThemeUseCaseImpl(themeRepository)
+        val factory = MainViewModel.getViewModelFactory(application, themeUseCaseImpl)
         viewModel = ViewModelProvider(this, factory).get(MainViewModel::class.java)
 
         findViewById<Button>(R.id.search).setOnClickListener {
